feat(pipeline): Refactor to workflow
