+++ cg_hermes/validate.py
@@ -2,8 +2,8 @@
 import logging
 from typing import FrozenSet
 
-from cg_hermes.config.pipelines import AnalysisType
 from cg_hermes.config.tags import COMMON_TAG_CATEGORIES
+from cg_hermes.config.workflows import AnalysisType
 from cg_hermes.constants.workflow import Workflow
 from cg_hermes.deliverables import Deliverables
 from cg_hermes.models.tags import TagMap
